Swap out `open` dependency for `webbrowser` (#86)
For graphviz debugging, open a browser window with `webbrowser` instead of `open`.

On macos, `open` calls out to a tool (also called `open`) which doesn't recognise URLs with dot files encoded in them as valid. Regardless of what program you tell it to open, it searches for a file (called "https://..."), doesn't find it, and returns an error.