Remove f64 (#377)
Spec was updated already in #358. I took out `LeafOp::Rzf64` too (the
only LeafOp that actually used it), also `SerSimpleType::F`. Added a
couple of other ClassicType instances for use in tests because we need
some (specifically, `ClassicType::Graph`s are a pain to construct,
cannot be `const` as they contain Boxes, and do not yet have ConstValue
instances)