Move resource validation routines to new module (#385)
This PR adds a new `resources::validate` module, and moves the different
resource logic bits from the Hugr validation there.

This is a tidying up change before adding new things for #362.

The only behaviour change is that `gather_resources` is ran on
initialization of the validation context, and the check for whether a
node has concrete input resources is done during the validation, when
calling `query_resources`.

Drive-by: Some missing docs for NodeType