Add a root NodeHandle type param to HugrView (#398)
I'll say this closes #380.

We could go further and add a type parameter in `Hugr<T: NodeHandle>`,
but that would probably be annoying to use and it'd break the pyo3
machinery.

Instead, most things should be interfacing via `HugrView`. #284 / #293
will move the builder references to views instead, so we will be able to
statically check the root type.