fix: subgraph boundaries with copies (#440)
Here I come again. The following test shows that the current
`SiblingSubgraph` implementation does not preserve function types, which
is a bit ugly. The reason for this is the existence of copy nodes at the
input boundary. Fixing this will probably also allow us to handle copies
at the output boundary, which are currently disallowed.

---------

Co-authored-by: Agustín Borgna <121866228+aborgna-q@users.noreply.github.com>
Co-authored-by: Seyon Sivarajah <seyon.sivarajah@quantinuum.com>
Co-authored-by: Alec Edgington <54802828+cqc-alec@users.noreply.github.com>
Co-authored-by: Craig Roy <croyzor@users.noreply.github.com>