fix: HugrView::get_function_type (#507)
Previously this was implemented separately for Hugr (correctly,
according to the doc), and for the other two HugrViews
SiblingGraph/DescendantsGraph (by deferring to the get_function_type of
the whole Hugr).

This is incorrect - the latter two should (surely?) return the function
type of the portion of the Hugr which they view.

However, we can both fix this and delete code, by moving the impl-Hugr
version to be the trait default, and removing the other two. See test
added (previously both were returning None).