Check ExtensionId strings well-formed (#491)
...as a dot-separated list of (dot-free) identifiers, adding a new type
`IdentList` of which `ExtensionId` is an alias.
E.g. no `foo..bar` or `.foo` or `5x` but `x5`, `foo.bar` ok.

`pub const fn new_unchecked` method allows bypassing the check. Nasty
but those constants are damn useful.

Also a `const_extension_ids` that automatically generates *tests* that
the names in the consts are valid.