feat!: SiblingSubgraph does not borrow Hugr (#515)
Removes HUGR borrows in `SiblingSubgraph`. This is necessary to be able
to use it to specify subgraphs for rewriting (as the rewriter will need
exclusive `mut` access to the Hugr).

I've taken the opportunity to clean up the interface by removing
close-to duplicate methods and making the naming more consistent.