feat: Validate TypeArgs to ExtensionOp (#509)
The TypeArgs might contain malformed things like "the type List<5>" - we
can only check that they make sense (i.e. that `5` is valid as an
argument for List) *given the ExtensionRegistry*, and nowhere outside
validation has the ExtensionRegistry, so we need to do this here.

Also rename `validate_children` (in ValidateOp) to
`validate_op_children)`, and `validate_operation` to
`validate_children`.