[refactor] resolve_extension_ops: take an ExtensionRegistry (#520)
This turns out to be pretty trivial, as ExtensionRegistry looks exactly
the same as HashMap as far as this code is concerned. Most of the effort
here is in fixing up imports and doclinks.

This addresses the first (easy) part of #508 but the second remains as
an issue/question for discussion, not attempting to resolve that here.