feat: Const nodes are built with some extension requirements (#527)
When using builder methods to put `Const` nodes on the graph, users must
pass an extra "provenance" argument which specifies which extension the
type of the constant value is coming from. The argument is an `Option`,
where `None` means that the constant is of a built-in type (like our
true and false predicates).

The `Const` nodes will be added to the graph with `input_extensions`
which are either the empty set or the singleton set of the const value's
provenance. Constant loading methods in the builder will make the
`LoadConst` node's `input_extensions` match those of the constant it's
connected too.

This also adds a constraint to extension inference that extension
requirements are the same on both ends of `Static` and `ControlFlow`
edges.

Resolves #525