Avoid need to rerun extension inference after outline-cfg rewrite (#534)
Mostly by updating the builder to allow specifying ExtensionSets in
places not previously possible: cfg_builder(), BlockBuilder::new, and
BlockBuilder::create.

I could go further: there is `fn block_builder` (and `fn
simple_block_builder` and others), which frustratingly take
`FunctionType`s (that I cannot update to Signature because Sig requires
an ExtensionSet not an Option thereof, so extra argument again)....

I guess requiring to specify `None` as extra param in the builder, may
be an OK price to keep the current behaviour of "run inference later"??

The above then avoids needing to `infer_and_validate` in OutlineCfg
tests after the rewrite, just `validate` works. It'd be good to keep the
other Inline/Outline rewrites and InsertIdentity like this too, but it's
probably not a desirable goal for anything taking a subgraph from the
user (i.e. (Simple/)Replace), where we should certainly allow to infer
requirements across the newly-inserted bit of graph...