fix: Make unification logic less strict (#538)
The `Plus` constraint has up till now been too strict in extension
unification. So far we have been deriving things like the following:
```
a := Plus("C", b)
a := {"A", "B", "C"}
implies
b := {"A", "B"}
```
(as in `minus_test`, which this PR removes)
However, plus should be meant as the union of the the singleton set that
it specifies with the extension set of another metavariable. Hence in
the above example the solution for `b` could be either `{"A", "B"}` or
`{"A", "B", "C"}`.

This means that if we have chains of circuit operations which all add
the "quantum" resource, they can all be represented as `Plus`
constraints