fix(Inference): Consider shunted metas when comparing equality (#555)
When comparing for equality of metavars, we need to consider that they
might have been merged to form a new meta, so
`UnificationContext::resolve` must be used. In this case, the variable
(LHS) is updated to the merged value in
`UnificationContext::merge_equal_metas` so we don't need to resolve that
one.