feat: Better subgraph verification errors (#587)
Added a `InvalidSubgraphBoundary` error variant so we get some info when
the subgraph verification fails.

Fixes:
- It seems that the check for boundary nodes having outside neighbours
(now `DisconnectedBoundaryPort`) was using an exhausted iterator, so it
was not running.

  With that fixed, now I get errors in the `portmatching` crate.
  
- `try_new_dataflow_subgraph` would return invalid subgraphs when the
input is directly connected to an output.
It now ignores the wires. I'll add an issue to discuss what should be
best behaviour in those cases.