Add RootChecked to guarantee root node type (#548)
* Factor out RootTagged as a subtrait of HugrView to allow separate
implementation. Most things can still just use HugrView (as it happens
everything that implements HugrView also implements RootTagged, but this
is not required).
    * HugrMut a subtrait of RootTagged not just HugrView
* Then add a new RootChecked struct storing any `AsRef<Hugr>` with a
fixed RootHandle.
* It allows `.as_ref()`, throwing away the extra information in the
RootHandle
     * If the underlying view is a HugrMut then it is too
* Use trait-default implementations of Hugr(View,(Mut)Internals), as the
latter check that `replace_op` does not break the bound on root-type
* But do not provide `.as_mut()` as that would allow bypassing and
invalidating the extra info in the RootHandle
* Like SiblingMut, check that nested instances only narrow the bound.
* Change the overridden-for-&(mut) Hugr impls of Hugr(Mut)Internals to
only work for `RootHandle=Node` to ensure the lack of checking in
replace_op there is safe.