feat: Builder and HugrMut add_op_xxx default to open extensions (#622)
A lot of this is refactoring existing code that explicitly uses
`add_node_xxx(NodeType::open_extensions(...))` to `add_op_xxx(...)`, and
updating tests - in most cases just `validate` -> `update_validate`.

Algorithmically, there was only one change - in `infer.rs` - where
Functions, FuncDefs and Aliases are given the empty ExtensionSet if they
are open (otherwise these end up unsolved).

Also there are some gymnastics in a couple of the tests where we want
the correct error out of validation but can't do inference because the
Hugr is, well, invalid....so we use a couple of different techniques to
carry solutions from earlier, valid, Hugrs over to the invalid one.
 
closes #424