feat: Add hashable `Angle` type to Quantum extension (#608)
After coding it up I have a couple of reservations about this:

* Calling the parameter "precision" suggests that it is an _approximate_
representation of a value, rather than an _exact_ representation of a
value that is restricted in its possibilities. I think that the latter
makes more sense, since "approximate values" are (I think) outside the
scope of HUGR, so another name might be better.
* Compilation passes that squash rotation gates using Euler-angle
decomposition will in general _not_ transform `Angle<N>` values into
`Angle<N>` values exactly. Therefore care will be needed in applying
such passes. Of course, even when applied to `float64` they are not
exact and can accumulate rounding error, so this isn't a new problem,
just one that's made more obvious.

The big advantage is that it makes angle values (and therefore circuits
containing them) equatable and hashable.

---------

Co-authored-by: Seyon Sivarajah <seyon.sivarajah@quantinuum.com>