Initial work on converting HUGR to a flake.
The use of devenv.yaml is retained, virtue of the flake-parts
approach outlined at `https://devenv.sh/guides/using-with-flake-parts/`.
devenv.lock and devenv.yaml are no longer needed, as the lock and
input source management is taken over by the flake.lock and flake.nix
respectively. I believe this is a better approach than bringing
devenv functionality directly into flake.nix (or, worse, duplicating
it).

The build of HUGR is provided as a derivation in default.nix. Building
succeeds, and testing passes, but I currently see no outputs in the result.

Cargo.lock was .gitignored, though this file is necessary when building
via Nix. Nix relies on this lock file to establish fixed-output
derivations (the alternative is manually specifying the dependencies and
their hashes within Nix, which is not a smooth development experience).
As such I have removed it from .git. I hope this does not prove to be
controversial.

I've added a .envrc which loads the dev env automatically on entering
the directory if authorised, and a couple of trusted substituters to
ease the initial download/build process. I've added .direnv to
.gitignore accordingly.