feat!: Drop the `pyo3` feature (#717)
Drop the remaining python binding definitions from the crate.

We currently re-wrap all objects in `tket2-py`, and only required this
for the `Into<PyErr>` implementations. As of
https://github.com/CQCL/tket2/pull/258, this is no longer needed.

The main benefit of removing this dependency is avoiding linking
problems when mismatching versions of `pyo3` are used by different
dependencies. The error is unavoidable due to it dynamically linking to
a single `python` lib, and cargo doesn't provide an easy way to deal
with it.