feat: Values (and hence Consts) know their extensions (#733)
* Make each Value able to report its ExtensionSet needed at runtime (and
hence, also each CustomConst)
* Also give OpTrait a method for the extension-delta. This is the delta
of the FunctionType, *for dataflow ops*, but can be defined for other
optypes too - specifically (here), Const ops and also Case. Use this in
both inference and for `NodeType::io_extensions()`
* Input extensions of every Const node can now be the empty set (i.e.
the default-for-ModuleOp `pure`), as the relevant extension will be
added in the delta - thus, drop separate extension parameter in builder
(add_constant, add_load_const, etc.). LoadConstant ops can now also have
an empty delta and open input-extensions as these can be figured out
from the Const. This is thus a step towards fixing #702...
* Also note slight issue in replace::test::cfg, pending #388
* Add `ExtensionSet::union_over(impl IntoIterator<Item=Self>) -> Self`
utility method

This should ease the way towards solving the rest of #702 and moreover
to removing `new_auto` - we should be able to constrain the
input-extensions for any ModuleOp to `pure` in inference and thus
*every* node can be created open rather than a mix, but those are all
for follow-up PRs.