feat: Minimal implementation for YAML extensions (#833)
This PR add a method `hugr::extensions::declarative::load_extensions`
that dynamically loads a set of extensions defined as YAML onto a
registry.

The code is mostly comprised of struct definitions to match the
human-readable serialisation format described in the spec, and some
methods to translate them into the internal hugr definitions.

There's a myriad of TODOs that should be addressed in future PRs,
including:
- Most parametric things (operations, type bounds, number of ports in a
signature, ...).
- Lowering functions, operations with non explicit signatures.
- Resolving the signature types.
- The syntax for describing these is not defined in the spec, so
currently there's just a couple of basic hard-coded types used for
testing: "Q" and "USize".
  
Here's an example of a supported definition:
```yaml
imports: [prelude]

extensions:
- name: SimpleExt
  types:
  - name: MyType
    description: A simple type with no parameters
  operations:
  - name: MyOperation
    description: A simple operation with no inputs nor outputs
    signature:
      inputs: []
      outputs: []
  - name: AnotherOperation
    description: An operation from 2 qubits to 2 qubits
    signature:
        inputs: [["Target", Q], ["Control", Q, 1]]
        outputs: [[null, Q, 2]]
```