ci: Skip invalid miri test, avoid concurrent undsoundness checks (#870)
- Fixes the failing unsoundness check in `main`
  https://github.com/CQCL/hugr/actions/runs/8158956607/job/22302072598
  
The test change in #852 moved the miri ignore flag, and it seems it has
to be placed after `rstest`'s cases.
  
- Adds a concurrency group for the unsoundness check.
The check currently takes around 15mins (and it will only grow as we add
more tests).
This makes it so old jobs get cancelled when pushing new changes to
main, and a new check is started.