feat!: Merge `Value` into `Const` (#881)
`Const` was a struct with a `Value` + a defined type. In practice, the
only `Value` that couldn't report its full type was `Sum`.

This PR merges the two structs, adding an explicit `sum_type` to the sum
variant.

Also:

- Removes `Type::check_type`, as this is now unnecessary. 
- Splits and clarifies the type check and constant errors;
`SumTypeError` is not a `ConstTypeError` is not a `CustomCheckFailure`.
  - Adds contexts to those errors, so we get better messages.
- `CustomConst` is now accessible from `::ops::custom` instead of
`::values`.

This change breaks the serialisation format.

Closes #874.

BREAKING CHANGE: Removed `Value`, merged into `Const`