feat: Add  a python package (#885)
Adds all the boilerplate for a `quantinuum-hugr` python package.

This includes conditional CI checks (dependant on the changed files),
and other metadata.

What's missing is a `release-please` configuration to generate python
changelogs. `release-plz` should already be configured to only include
changes that modify the rust project.

This is the foundation required for #486 and
https://github.com/CQCL/guppylang/issues/173, and it should make adding
another package for #866 quite easy.