feat: Pydantic model for the serialization schema (#888)
This PR ports the pydantic definition from guppy (including
https://github.com/CQCL/guppylang/pull/172), and updates it to sync up
with the current schema.

Changes to the schema:

- Renamed the `InputExtensions` class to `ExtensionSet`
- Added a named definition for `Const`
- Dropped `DummyOp`. It was referenced in a couple field types, but the
model was never defined.

The main difference with the model in `guppy` is the use of
`RootModel`-based classes instead of variables assigned to a
`TypeAliasType` for defining aliases on type unions.
They both work the same, but `mypy` doesn't like using variables in
types.

Additionally, I updated some READMEs and added a CI check to ensure that
the json in `specification/schema/` is always up to date.