chore: Configure pre-commit, update the justfile (#890)
Setups `pre-commit` instead of the ad-hoc `.github/pre-commit` file, and
cleans up the `just` recipes.

The main benefit of this is to avoid running both codebase tests on each
commit, since `pre-commit` filters the required checks based on the
modified files.

Updates `DEVELOPMENT.md` with the new commands. It should be a bit more
straightforward now.

drive-by: Fixes for the `trailing-whitespace` and `end-of-file-fixer`
hooks.