chore: release v0.3.0-alpha.1 (#916)
With the rename of the crate from `hugr_quantinuum` to `hugr` (#914) we
broke `release-plz`' automatic changelog generation (since it cannot
find the crate anymore), and all links to crate-related urls.

This alpha release should fix those issues, in addition to letting us
claim the new crate name.

I added the partial changelog bits from #898, #893, and #840 since they
won't be included in the new package's PRs.

drive-by: Move CHANGELOG and README from the root to the `hugr` dir to
be more consistent with other packages in the workspace, and symlink
them instead.