feat!: Flatten `LeafOp` (#922)
Moves the variants of `LeafOp` into `OpType`. Closes #817.

It seems `Lift` was missing from the schema, and in its place there was
a `TypeApplication` struct. I replaced it with lift.

As an aside, the `CustomOp -> ExternalOp -> {ExtensionOp, OpaqueOp}`
chain seems unnecessary. We should be able to merge `CustomOp` and
`ExternalOp` into one, and get rid of multiple `as_ref`s in the code.
This doesn't modify the schema, so I'll push it as a separate PR.

---------

Co-authored-by: Seyon Sivarajah <seyon.sivarajah@quantinuum.com>