ci: fix package name in release-please (#930)
Fixes some config in release-please so now it looks for the right
version tags `hugr-py-v*.*.*`. It doesn't seem to like alphas though, so
the first releases will need some manual tweaking (it tries to do 0.1.0
otherwise).

The CI job is failing with some github permission error... I copied the
guppy one to try and fix that, but I can run it locally via the CLI if
needed for these first releases.