feat: 'Replace' rewrite returns node map (#929)
This is a preliminary step towards using merging basic blocks, where the
Replace rewrite needs to be used as part of a larger transformation -
this makes it possible to use compositionally.

Along the way, refactor the tests a bit (changing `verify_apply` to
`check_same_error` and doing all the bad/`Err`-returning cases after the
success case).