feat: Encoder metadata in serialized hugr (#955)
Closes #950.

Adds a `hugr-rs v0.3.0` or a `hugr-py v0.1.0` label to the serialialised
hugrs under the optional `encoder` field.
This should help with debuggability and error reporting in the future,
for now we just ensure that we start adding the labels. The added field
to the schema is backwards-compatible.

I had to add a `__version__` field at the root of the python package.
This is automatically updated by `release-please`, so it shouldn't be
much of an issue.

drive-by: force `poetry update` before generating the schema in `just
update-schema`, to make sure we are using the same pinned pydantic
version as CI.