feat!: bring back Value (#967)
This change is primarily motivated by the serialisation schema. With
Const and Value merged the recursive Const type was an Op , so was
serialised with "op" and "parent" fields. These fields were only
relevant for "root" Consts.

Now, Const is an OpType, while Value is not. 

An unrelated change is included: The serialisation schema for SumType
now consists of the correct list of TypeRows.

Note that serialisation still does not match the schema, in particular
SumTypes are still not properly handled.

---------

Co-authored-by: Agustín Borgna <121866228+aborgna-q@users.noreply.github.com>