feat: merge basic blocks algorithm (#956)
Any basic block with a single successor that has no other predecessors
can be merged. (So long as the successor is not the exit block; nor the
entry block, as that implicitly has another predecessor).

Implement via Replace (2 BBs - > one BB containing 2 DFGs, which adopt
the original BB's children) and 2*InlineDFG.

Closes #561.