test: proptests for round trip serialisation of `Type`s and `Op`s. (#981)
Note that I make no attempt to generate "valid" types. In particular
this means that type variables will be grossly inconsistent. It would be
possible to improve this by passing a type environment through all the
arbitrary instances in `Parameters`, although it's not clear to me just
how far that would take us.

We can't use `prop_recursive` for mutually recursive types, so I have
reimplemented it using `Parameters`.