ci: publish to pypi on releases or workflow_dispatch for tags (#1068)
Fixes merging release-please PRs not triggering the pypi release
workflow.
This is due to workflows not triggering other workflows when using the
default `GITHUB_TOKEN`.
  
The pypi release workflow ran when a tag matching `hugr-py-v**` was
pushed by anyone. This was a bit dangerous and prone to fat-fingering.
Now the workflow is triggered either by new releases with a matching tag
(as created by release-please), or by manually triggering the workflow
on a tag (so we can still trigger it if needed).
This should be harder to trigger by accident.