ci: Enable crates.io auto-release, but require certain people to approve the PRs (#1069)
This setups a restricted ownership over the changelog files, so the
release PRs by `release-please` or `release-plz` have to be approved by
a restricted set of people.

We can then enable auto-releases for rust (for python is already
enabled).

Feel free to suggest other users to add to the list.