feat!: Move passes from `algorithms` into a separate crate (#1100)
Closes #1000 .

Also:
- Moved some tests that were making use of these passes (or their test
code) from the `hugr` crate into the `hugr-passes` crate.
- Moved some functions from test code into the `utils` module and made
it public.
- Made `Hugr::new()` public. I thought I could avoid this by using the
`CFGBuilder` but doing so caused some extension-inference tests to fail,
for reasons I couldn't fathom.

Do we want to update the `release-plz` configuration? Not sure exactly
what to do here.

BREAKING CHANGE: The `algorithms` module is removed from `hugr`. It's
functions are now in a separate `hugr-passes` crate.

---------

Co-authored-by: Agustín Borgna <121866228+aborgna-q@users.noreply.github.com>