refactor: Add a `hugr-core` crate (#1108)
Moves all the code from `hugr` into a `hugr-core` subcrate.
This way, `hugr-passes` (and any other new subcrate) can depend on the
core definitions and get reexported in `hugr`.

This PR does not. Change any definition or visibility in the code, as
the renamed files are noisy enough.
In particular, having `hugr-core` will let us unseal `HugrInternals` and
`HugrMutInternals` so they can be used in `tket2`, but that's work for
another PR.

As a bonus, #1100 is no longer a breaking change since we can re-export
the library without a cyclic dependency.

Closes #294