refactor!: FunctionBuilder takes impl Into<PolyFuncType> (#1220)
Currently FunctionBuilder::new and the `define_function` method both
take a PolyFuncType; this PR changes that to a more-general `impl
Into<PolyFuncType>`.

BREAKING CHANGE: clients calling these methods with a FunctionType's
`.into()` will fail to compile with a "type annotations required" error,
which can be fixed by deleting the `.into()`.