fix!: Allow Const and FuncDecl as children of Modules, Dataflow Parents, and CFG nodes (#46)
We take the opportunity to fix a bug where the promise to ignore
duplicate calls to `emit_func` was not upheld, and to clean up the now
unnecessary `Emission`.

Fixes #4 , although we go further than required by this issue.

BREAKING CHANGE: rename `emit_global` to `emit_func`. Replace `Emission`
type with `FatNode<FuncDefn>`.