fix: Panic on `SimpleReplace` with multiports (#1324)
Fixes #1323

This fix generalises the one from #1191.
SimpleReplace was too eager in disconnecting/connecting edges to the
replacement graph, and that caused issues when querying the neighbours
of multiports.

This gets resolved by delaying all new connections to the replacement
until after we have computed all of them.
We don't need to do explicit disconnections to the replaced subgraph,
since the nodes get removed anyway.

---------

Co-authored-by: Douglas Wilson <141026920+doug-q@users.noreply.github.com>