fix: Sibling extension panics while computing signature with non-dataflow nodes (#1350)
The signature computation for a `SiblingSubgraph` takes the union of the
nodes' extensions. This didn't contemplate non-dataflow nodes like
constants, and caused a runtime panic if one was present.

Most of the diff is adding a constant node in the tests. 

This is a fix for https://github.com/CQCL/tket2/issues/507