refactor!: add `LLVMSumValue` (#63)
The payoff here is in the simplification of lowering of `CFG` and
`Conditional`. By factoring out the destructuring logic, we will be able
to reuse this in other places. In particular generating destructors.
This has the unfortunate side-effect of changing the test case output.
The price one pays for these golden tests.

BREAKING CHANGE: some types in public interfaces have changed slightly,
in particular variant tags are now consistently `usize`