ci: Reduce rust coverage report delay (#1370)
CI was running tests on `{MSRV, beta, nightly}` first, and only then
running tests with coverage instrumentation on nightly.

Since those tests are not on the critical path for the required checks,
we can run the nightly tests directly with instrumentation on the first
pass, shaving ~2mins from the commit-to-codecov report delay.

I'm only doing this for rust, as the python tests _are_ in the required
checks sets and we don't want to slow down those.

drive-by: The python tests job had the same name as the linting job.

---------

Co-authored-by: Douglas Wilson <141026920+doug-q@users.noreply.github.com>