feat!: Serialised extensions (#1371)
Fear not the diff, mostly schema update.


- Define a Pydantic model and serialised schema for extensions.
- Update the rust `SignatureFunc` serialisation to be compatible with
this.
- Serialized extension "declarations" can state they require binary
compute or validation functions. `SignatureFunc` reports this if they
are missing, and it is up to the caller (typically validation) to decide
how to recover from this.

Closes #1360 
Closes #1361

Addresses parts of #1228 

Best reviewed as individual commits.

BREAKING CHANGE: `TypeDefBound` uses struct-variants for serialization.
`SignatureFunc` now has variants for missing binary functions, and
serializes in to a new format that indicates expected binaries.