ci: Fix pr-title not running (#1384)
It seems that skipping reusable workflow jobs causes inconsistent check
names between `pull_request_target` and `merge_group`.
For the non-skipped PR runs we get a `local_job/remote_job` name,
but the skipped jobs in the merge queue are just called `local_job`....
(even if skipped in the called workflow instead of locally).

I updated the reusable workflow to skip individual steps instead of the
whole job, so now we shouldn't have to deal with that inconsistency.

I will need to update the required check name to `local/remote`