feat(hugr-py)!: user facing Extension class (#1413)
Allows users to define extensions in Python, attaching operation, type
and value definitions to them. Also provides utilities for then using
those objects when building HUGRs.

Diff is big, but a lot of it is poetry and schema changes. Open to
suggestions as to how to break up but I think it is fairly
self-contained.

Closes #1374
Closes #1412 
Also as a drive-by adds the missing `binary` field to serialised opdef
signatures.


BREAKING CHANGE: `AsCustomOp` replaced with `AsExtOp`, so all such
operations now need to be attached to an extension.

---------

Co-authored-by: Alec Edgington <54802828+cqc-alec@users.noreply.github.com>