refactor!: flatten `CustomOp` in to `OpType` (#1429)
In new workflow most of the time we just want to work with extension
ops, and opaque are just used for serialisation. To that end flatten
`CustomOp` out. If there are `OpaqueOp`s in the HUGR it means the
extension was not available to resolve in to. Will be disallowed in
follow up PR.

Relates to #1362 


BREAKING CHANGE: `CustomOp` removed, `OpType` now contains `ExtensionOp`
and `OpaqueOp` directly. `CustomOpError` renamed to`OpaqueOpError`.