chore: Use local `ruff` and `mypy` for pre-commit (#1440)
Ensures that the checks run by `pre-commit` are the same as the ones run
in CI.
Before, we used pre-commit actions with their own versions that needed
to be updated separately and tended to get out of date.

As an example, the ruff update errors from #1437 didn't appear locally
when running `just check`.

drive-by: Update `ruff` to latest version