feat: Bring in the pure-python renderer from guppy (#1462)
Closes #1407.

Ports @mark-koch's rendering code from guppy, since it got deleted from
there in the last hugr builder update.
Adds a `Hugr.render_dot` and a `Hugr.store_dot` method.

I'm using `syrupy`—a snapshot pytest extension—for testing it. I'm not
sure if we want to be this strict with the generated output though.