feat(hugr-py)!: pretty printing for ops and types (#1482)
Closes #1468 
Closes #1470

Previous testing wasn't covering label formatting because it requires
actually rendering the dot in to an image. Have enabled that, but put it
behind an env var because it significantly slows down tests.

BREAKING CHANGE: rename `Custom.name` to `Custom.op_name` and
`Func(Defn/Decl).name` to `f_name` to allow for new `name` method