refactor!: rework extensions interface (#119)
`CodegenExtension` is replaced with a new trait of the same name, that
instead of defining extension behaviour provides a convenient interface
for registering a set of extensions. The `prelude` extension gives the
best example, exercising all the features.

Extensions are now registered via methods on `CodegenExtsBuilder`, each
method takes a callback that is stored in the builder. The builder is
consumed by `finish` which returns a collection of "Callback Maps". One
of the big wins here is that the "Type Callback Map" does not depend on
a `H: HugrView` parameter.

The "root" `&'c Context` now lives in `EmitModuleContext`, as well as in
`TypingSession`.

`TypeConverter` is now the "Type Callback Map". This is implemented
using the new `src/utils/type_map.rs`, which we intend to leverage for
computing "def-hooks" and "destructors" for types.
 
We go to some effort to ensure the callbacks are not required to have
`'static` lifetimes, nothing inside `crate::custom` assumes or requires
this. However, we do in several places (e.g. `TypingSession`,
`EmitModuleContext`) fix these lifetimes to `'static`, in lieu of adding
additional lifetime parameters to these types.

We have removed the `guppy.rs` test case, which was ignored, rather than
rework it to use the new interface.
Otherwise, test cases are largely unchanged, and there are no snapshot
changes.

All extensions have been changed to use the new interface, which is
largely just moving code around. In some cases error messages are
consolidated by the new tools providing more appropriate inputs to
extension points. See for example `CodegenExtsBuilder::custom_const`.


At several points we mark TODO refactorings that will be carried out in
the near future.

BREAKING CHANGE: extensions interface in `src/custom.rs` entirely
reworked.

---------

Co-authored-by: Agustín Borgna <121866228+aborgna-q@users.noreply.github.com>