feat: Binary serialisation format for hugr-model based on capnproto. (#1557)
A binary serialisation format for `hugr-model` based on capnproto.

- Export from `hugr-core` to `hugr-model` now deduplicates terms on the
fly.
- Removed types on ports and instead used the signature. This allows for
more deduplication and also provides a natural place for extension set
annotations which was missing before.
- Added the `NodeId` of the node that defines a parameter to the indexed
variant of a `LocalRef` so that deduplication does not accidentally
merge two local variables that refer to different parameters.

Serializing the same 2000 2-qubit gate + 1000 1-qubit gate circuit by
going from `core` via the `model` takes 5.5682 ms per roundtrip on Apple
M3. Meanwhile the JSON roundtrip baseline is 10.760 ms. A significant
portion of the time is spent on needless copies of signatures since
`export` and `import` have to interact with the `hugr-core` API (#1551),
so there is still a lot of potential to make this faster yet.